change how admin authorization works #170

+++ backend/BoomaEcommerce.Services/Users/SecuredUserService.cs
@@ -78,8 +78,10 @@ namespace BoomaEcommerce.Services.Users
 
             var shoppingCart = await _next.GetShoppingCartAsync(userGuidInToken);
 
-            return await _next.AddPurchaseProductToShoppingBasketAsync(shoppingBasketGuid, purchaseProduct);
-
+            if (shoppingCart.Baskets.FirstOrDefault(basket => basket.Guid == shoppingBasketGuid) != null)
+            {
+                return await _next.AddPurchaseProductToShoppingBasketAsync(shoppingBasketGuid, purchaseProduct);
+            }
             throw new UnAuthorizedException($"User {userGuidInToken} cannot add a purchase product to {shoppingBasketGuid} shopping basket.");
         }
 
@@ -90,8 +92,10 @@ namespace BoomaEcommerce.Services.Users
 
             var shoppingCart = await _next.GetShoppingCartAsync(userGuidInToken);
 
-            return await _next.DeletePurchaseProductFromShoppingBasketAsync(shoppingBasketGuid, purchaseProductGuid);
-
+            if (shoppingCart.Baskets.FirstOrDefault(basket => basket.Guid == shoppingBasketGuid) != null)
+            {
+                return await _next.DeletePurchaseProductFromShoppingBasketAsync(shoppingBasketGuid, purchaseProductGuid);
+            }
 
             throw new UnAuthorizedException($"User {userGuidInToken} cannot remove a purchase product from {shoppingBasketGuid} shopping basket.");
         }
@@ -104,9 +108,10 @@ namespace BoomaEcommerce.Services.Users
 
             var shoppingCart = await _next.GetShoppingCartAsync(userGuidInToken);
 
-         
-            return await _next.DeleteShoppingBasketAsync(shoppingBasketGuid);
-
+            if (shoppingCart.Baskets.FirstOrDefault(basket => basket.Guid == shoppingBasketGuid) != null)
+            {
+                return await _next.DeleteShoppingBasketAsync(shoppingBasketGuid);
+            }
 
             throw new UnAuthorizedException($"User {userGuidInToken} can not delete a basket from {shoppingCart.Guid} shopping cart.");
         }
